Update issue 226 EFLIB:   Add macro for defining weak_ptr. SALVIA:   Add functions to shader_reflection.   Add vx_shader_unit to shader_unit.h   Move get_stream_desc to stream_assembler.   Remove some warnings. SASL:   Add sasl_create_ia_shim to drivers.   Remove used dependencies from host_impl.h.   Rename member functions and member variable in reflection_impl.   Implements ia_shim, interp_shim and om_shim (not completed).   Add shims project to sasl.   Remove 'lang' field from reflection_impl because 'module_sem_' contains language.   Remove cg_jit.h and cg_jit.cpp.   Link sasl_drivers with sasl_shims.   Add APIs to drivers_api.h to support shim creation.   Improve implementation of host_impl.   Add shader_unit_impl to host project.

+++ salviax/src/resource/mesh/sa/mesh_io_collada.cpp
@@ -269,7 +269,7 @@ vector<h_mesh> build_mesh( dae_mesh_ptr m, skin_info* skinfo, renderer* render )
 		vector< vector<char> >			buffers_data(skinfo ? vertex_attributes_count + 2 : vertex_attributes_count ); // If skinfo is available, last two buffers for joints and weights.
 		vector<uint32_t>				attribute_merged_indexes;
 		index_dict_t					index_dict; // a map from grouped index to merged index.
-		size_t							merged_vertex_counter = 0;
+		uint32_t						merged_vertex_counter = 0;
 		indexes_of_vertex_attributes	index_group(vertex_attributes_count);
 
 		// FOR: index of triangle vertexes.
